move all worker related logic prom profiler.py to worker_manager_mixin.py
